Rename cli.py
