+++ src/gpkg/__init__.py
@@ -1,5 +1,4 @@
 from gpkg._api import fetch, install, show, upgrade
-from gpkg._cli import cli
 from gpkg._models import PackageInfo, Status, StorageType
 from gpkg._storage import Storage
 
@@ -8,7 +7,6 @@ __all__ = [
     "install",
     "show",
     "upgrade",
-    "cli",
     "PackageInfo",
     "Status",
     "StorageType",
