+++ src/gpkg/__main__.py
@@ -1,3 +1,3 @@
-from gpkg import cli
+from gpkg.cli import cli
 
 cli()
